Add PlanningExplorer scrapers:
    * Blackburn
    * Camden - done (has OSGB)
    * Charnwood (done, but site has a problem with its info links)
    * Crewe and Nantwich
    * East Staffordshire
    * Epping Forest
    * Hackney - done
    * Kennet
    * Lincoln (also has WAM)
    * Liverpool
    * South Shropshire
    * South Tyneside
    * Stockport
    * Tamworth

I've included Camden, as the scraper works (their site is down today).
The scraper for Charnwood is fine (both the search results page and the info links fail to display in a browser, but 
we are getting all the info out fine, and the links are correct. This appears to be a missing template on their 
site).

I've included several authorities for which the scrapes work fine, but which don't have any postcodes. Some of these 
have OSGB coords for each app, which we could maybe use instead.

    * Broadland (no postcodes)
    * Forest Heath (no postcodes)
    * South Norfolk (appears to have WAM as well) (No postcodes - but has OSGB coords)
    * Trafford (no postcodes)
    * West Oxfordshire (no postcodes)

There are a few PlanningExplorer sites that are currently down, and which I haven't been able to do yet:

    * Merton (site down)
    * Shrewsbury (site is broken)
    * Swansea (site is broken)
    * Waltham Forest (currently giving an error)